working on update 2
Started minimizing number of database connections needed for getting
information. Removed name requirement. If no name entered, then it will
be department, if a name is entered in only one field error is thrown.
otherwise it proceeds as normal. Setup script was formatted, and test
script is background work for the minimizing connections statement